Add react components to logically layout website content architecture
Created dummy (empty) components to layout the architecture of the
website.
Documented layout to 'README.md' (very succintly)
Updated 'App.tsx' to render the components instead of the starter
vite app and deleted the default svgs that are unrelated to the
project.
A 'shared' directory was created for components that will be imported
by various components, like the company logo.