Add 'colorStyle' property to Logo to choose correct color scheme
Applicatioin needs to be able to render the logo in two different
color schemes, but current implementation lacked a way to select
between the two.
Added 'colorStyle' property to solve this issue. Possible values
'light' and 'dark' choose between the two color schemes. The prop
refers to the color of the background